sums of lengths should use size_t.
git-svn-id: https://svn.code.sf.net/p/bzflag/code/trunk/bzflag@17129 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ src/bzfs/bzfsMessages.cxx
@@ -1063,7 +1063,7 @@ void PackVars::endMessage()
 void PackVars::sendPackVars(const std::string &key)
 {
   std::string value = BZDB.get(key);
-  int pairLen = key.length() + value.length() + 2 * sizeof(uint32_t);
+  size_t pairLen = key.length() + value.length() + 2 * sizeof(uint32_t);
   if ((pairLen + msg->size()) > MaxPacketLen) {
     endMessage();
     startMessage();
